% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_themes.R
\name{theme_Tim}
\alias{theme_Tim}
\title{A \link{ggplot2} theme for Isabella Benabaye's personal use}
\usage{
theme_Tim(
  title_family = "Source Sans Pro",
  text_family = "Source Sans Pro",
  base_size = 7,
  base_size_pane = 13,
  text_color = "black",
  bg_color = "#ffffff",
  line_color = "black",
  plot_margin = margin(5, 5, 5, 5),
  legend_margin = margin(-15, -10, -15, -15),
  plots_pane = FALSE,
  md = FALSE
)
}
\arguments{
\item{title_family}{title elements font family}

\item{text_family}{text elements font family}

\item{base_size}{base font size}

\item{text_color}{color of the text & title elements}

\item{bg_color}{plot background color}

\item{line_color}{color of line elements}

\item{plot_margin}{plot margin specifications}

\item{plots_pane}{indicate whether plots will be viewed in plots pane or not}

\item{md}{indicate whether to use markdown elements for text}
}
\description{
\code{theme_Tim} is a \link{ggplot2} theme in Tim Monko's style - courtesy of inspiration from Isabella Benabaye's Ib_theme and Ib package. The theme was specifically designed to be used for plots in manuscripts. Defaulting to a size suitable for resizing without having to worry about relative sizing and font sizing.
}
\details{
By default (\code{plots_pane = FALSE}), the theme adjusts the text sizes for
printing.\code{plots_pane = TRUE} is meant to be used when viewing plots in the
plots pane and text sizes are not adjusted.  There is also an option (\code{md = TRUE}) to use markdown theme elements from \code{ggtext} instead of
\code{element_text()}.
}
